// src/components/drone/enum/drone.enum.ts
export enum DroneState {
  IDLE = 'idle',
  LOADING = 'loading',
  LOADED = 'loaded',
  DELIVERING = 'delivering',
  DELIVERED = 'delivered',
  RETURNING = 'returning',
}

export enum DroneModel {
    LIGHTWEIGHT = 'LightWeight',
    MIDDLEWEIGHT = 'MiddleWeight',
    CRUISERWEIGHT = 'CruiserWeight',
    HEAVYWEIGHT = 'HeavyWeight',
}